requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask_marshmallow==1.2.1
marshmallow==3.21.1
python-dotenv==1.0.1
flask-cors==4.0.0
flask-restx
barcode
python-multipart==0.0.9
python-docx==1.1.0
pandas==2.2.2
reportlab==4.2.5
qrcode==7.4.2
flask-mail==0.9.1
pyboleto==0.3.1

pytest==8.2.2
Flasgger==0.9.5
marshmallow-sqlalchemy==0.29.0